Debugging Rad4_error.cpp
For some reason, being told in VS that trying to access null/undeclared
value. Occurs around line 140, when we try and return a value from the
tiles array.  Not sure what is going on with this problem. currently
printing out values being calculated/tried to access. Seems to be having
trouble acccessing tile[0] .